refactor controller traits to be internal
