+++ crates/libcgroups/src/v2/controller.rs
@@ -3,6 +3,6 @@ use std::path::Path;
 
 use crate::common::ControllerOpt;
 
-pub trait Controller {
+pub(super) trait Controller {
     fn apply(controller_opt: &ControllerOpt, cgroup_path: &Path) -> Result<()>;
 }
